+++ app/src/main/java/com/starters/hsge/data/interfaces/UserInfoGetInterface.kt
@@ -1,0 +1,10 @@
+package com.starters.hsge.data.interfaces
+
+import com.starters.hsge.data.model.remote.response.UserInfoGetResponse
+
+interface UserInfoGetInterface {
+
+    fun onGetUserInfoSuccess(userInfoGetResponse: UserInfoGetResponse ,isSuccess: Boolean, code: Int)
+
+    fun onGetUserInfoFailure(message: String)
+}
