+++ app/src/main/java/com/starters/hsge/presentation/main/mypage/MyPageFragment.kt
@@ -6,7 +6,10 @@ import android.view.View
 import androidx.navigation.fragment.findNavController
 import com.starters.hsge.R
 import com.starters.hsge.data.api.UserInfoGetApi
+import com.starters.hsge.data.interfaces.UserInfoGetInterface
 import com.starters.hsge.data.model.remote.response.UserInfoGetResponse
+import com.starters.hsge.data.service.NicknameService
+import com.starters.hsge.data.service.UserInfoGetService
 import com.starters.hsge.databinding.FragmentMyPageBinding
 import com.starters.hsge.network.RetrofitClient
 import com.starters.hsge.presentation.common.base.BaseFragment
@@ -15,7 +18,7 @@ import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class MyPageFragment : BaseFragment<FragmentMyPageBinding>(R.layout.fragment_my_page) {
+class MyPageFragment : BaseFragment<FragmentMyPageBinding>(R.layout.fragment_my_page), UserInfoGetInterface {
 
     var profileImage : Int = 0
     var nickname : String = ""
@@ -29,7 +32,7 @@ class MyPageFragment : BaseFragment<FragmentMyPageBinding>(R.layout.fragment_my_
         super.onViewCreated(view, savedInstanceState)
 
         // 사용자 정보 조회
-        tryGetUsersInfo()
+        UserInfoGetService(this).tryGetUserInfo()
 
         binding.ivSettings.setOnClickListener{
             findNavController().navigate(R.id.action_myPageFragment_to_settingsFragment)
@@ -66,48 +69,34 @@ class MyPageFragment : BaseFragment<FragmentMyPageBinding>(R.layout.fragment_my_
         }
     }
 
-    private fun goneBtmNav(){
-        (activity as MainActivity).binding.navigationMain.visibility = View.GONE
+    private fun goneBtmNav(){ (activity as MainActivity).binding.navigationMain.visibility = View.GONE }
+
+    override fun onGetUserInfoSuccess(userInfoGetResponse: UserInfoGetResponse, isSuccess: Boolean, code: Int) {
+        if (isSuccess) {
+            // 사용자 정보 가져오기
+            profileImage = userInfoGetResponse.profilePath
+            nickname = userInfoGetResponse.nickname
+            town = userInfoGetResponse.town
+            latitude = userInfoGetResponse.latitude
+            longitude = userInfoGetResponse.longtitude
+            radius = userInfoGetResponse.radius
+
+            var townStrList =  town.split(" ")
+            if (townStrList.size == 3) {
+                splitTown = townStrList[1] + " " + townStrList[2]
+            } else if (townStrList.size > 3) {
+                splitTown = townStrList[1] + " " + townStrList[2] + " " + townStrList[3]
+            }
+
+            // 사용자 정보 적용
+            binding.ivUserProfile.setImageResource(profileImage)
+            binding.tvUserNickname.text = nickname
+            binding.tvUserLocation.text = splitTown
+        }
     }
 
-    private fun tryGetUsersInfo(){
-        val userInfoGetApi = RetrofitClient.sRetrofit.create(UserInfoGetApi::class.java)
-        userInfoGetApi.getUserInfo().enqueue(object :
-            Callback<UserInfoGetResponse> {
-            override fun onResponse(
-                call: Call<UserInfoGetResponse>,
-                response: Response<UserInfoGetResponse>
-            ) {
-                if (response.isSuccessful) {
-                    val result = response.body() as UserInfoGetResponse
-
-                    // 사용자 정보 가져오기
-                    profileImage = result.profilePath
-                    nickname = result.nickname
-                    town = result.town
-                    latitude = result.latitude
-                    longitude = result.longtitude
-                    radius = result.radius
-
-                    var townStrList =  town.split(" ")
-                    if (townStrList.size == 3) {
-                        splitTown = townStrList[1] + " " + townStrList[2]
-                    } else if (townStrList.size > 3) {
-                        splitTown = townStrList[1] + " " + townStrList[2] + " " + townStrList[3]
-                    }
-
-                    // 사용자 정보 적용
-                    binding.ivUserProfile.setImageResource(profileImage)
-                    binding.tvUserNickname.text = nickname
-                    binding.tvUserLocation.text = splitTown
-                }
-            }
-
-            override fun onFailure(call: Call<UserInfoGetResponse>, t: Throwable) {
-                Log.d("MyPage 실패", t.message ?: "통신오류")
-
-            }
-        })
+    override fun onGetUserInfoFailure(message: String) {
+        Log.d("UserInfoGet 오류", "오류: $message")
     }
 
 }
